pyproject.toml
[project]
name = "moviepicker"
version = "0.1.0"
description = ""
authors = [
    {name = "Nawal",email = "n.ahmednazim@uqconnect.edu.au"}
]
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "boto3 (>=1.38.21,<2.0.0)",
    "google-auth (>=2.40.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "flask-cors (>=6.0.0,<7.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
